requirements.txt
celery==3.1.23
dj-database-url==0.4.1
Django==1.9.7
Jinja2==2.8
MarkupSafe==0.23
Unidecode==0.4.18
diff-match-patch==20121119
django-bower==5.1.0
django-cachalot==1.2.1
django-compressor==2.0
django-daterange-filter==1.1.1
django-debug-toolbar==1.4
django-extensions==1.6.7
django-geoposition==0.2.2
django-htmlmin==0.9.1
django-import-export==0.4.5
django-localflavor==1.3
django-redis==4.4.3
django-salmonella==1.0
django-scheduler==0.8
django-selectable==0.9.0
django-selectable-filter==0.1.5
django-smart-selects==1.1.1
django-spaghetti-and-meatballs==0.2.0
django-su==0.5.2
django-suit==0.2.19
django-suit-locale==1.0.10
django-suit-redactor==0.0.4
docutils==0.12
flower==0.9.1
gunicorn==19.6.0
hiredis==0.2.0
mysqlclient==1.3.6
numpy==1.11.0
pandas==0.18.1
Pillow==3.2.0
psycopg2==2.6.2
python-slugify==1.1.4
pytz==2016.4
six==1.10.0
sorl-thumbnail==12.4a1
tablib==0.11.2
wiki==0.1.2